fix delete analyist problem

+++ src/main/java/com/caudbdesign/dbTeamProject/Admin/Analystinfo/AnalystController.java
@@ -4,6 +4,7 @@ package com.caudbdesign.dbTeamProject.Admin.Analystinfo;
 import com.caudbdesign.dbTeamProject.Item.Stock.AnalystInfo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -18,8 +19,10 @@ public class AnalystController {
 
   @PostMapping("/admin/deleteAnalyst")
   @CrossOrigin
-  public void deleteAnalyst(@RequestBody AnalystInfo info) {
+  public ResponseEntity<?> deleteAnalyst(@RequestBody AnalystInfo info) {
+    if(!analystService.isAnalystExist(info.getAnalyst_id())) return ResponseEntity.badRequest().build();
     analystService.deleteAnalyst(info.getAnalyst_id());
+    return ResponseEntity.ok().build();
   }
